pyproject.toml
[project]
name = "ethos"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.3",
    "langchain>=0.3",
    "langchain-anthropic",
    "langchain-openai",
    "langchain-community",
    "tavily-python",
    "python-dotenv",
    "pyyaml",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
daytona = [
    "daytona",
]
dev = [
    "pytest",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ethos"]
